清理没用的东西

+++ Link22/Utils/MsgUtil.cpp
@@ -50,7 +50,7 @@ std::string MsgUtil::CharArrayToBitStr(const uint8_t *data, int char_length) {
 std::string MsgUtil::StrToBitStr(const std::string &data) {
     size_t len = data.length();
     std::string res;
-    for (int i = 0; i < len; i++) {
+    for (size_t i = 0; i < len; i++) {
         std::bitset<8> bits = std::bitset<8>(data[i]);
         res += bits.to_string();
     }
@@ -66,9 +66,9 @@ std::string MsgUtil::BitStrToStr(const std::string &data) {
     if (data.length() == 0 || data.empty()) return "";
     size_t len = data.length();
     std::string res;
-    for (int i = 0; i < len; i = i + 8) {
+    for (size_t i = 0; i < len; i = i + 8) {
         std::string tmp;
-        for (int j = 0; j < 8; j++) {
+        for (size_t j = 0; j < 8; j++) {
             tmp += data[i + j];
         }
         std::bitset<8> ch = std::bitset<8>(tmp);
